Changed "properties" / "contents" accessors
Now using the name "properties" for the full contents of the document/revision, with "userProperties" for the non-underscored properties.
CouchDocument.properties thus now returns all properties including _rev. This is important, so that when a client modifies this dictionary and calls putProperties: with it, the _rev key is still there. Otherwise we can't detect conflicts properly because we wouldn't know which revision of the document this was based on.