Fixed nasty reentrancy bug in RESTOperation.
It was possible for an async RESTOperation to complete while another one was blocked in -wait. This could cause problems in client code due to reentrant calls -- for example, a table view might be blocked drawing a cell waiting for a -GET method to return a document, and while it's waiting it would get a completion call for a query result, and decide it needed to redraw the table … badness ensues.
Fixed this by having -completedWithError: detect if it's invoked during a different operation's -wait call, and if so defer itself using a delayed perform.