Fixes to compile with 10.7 SDK
Careful with those macros; if we're building a Mac app, the IPHONE_OS
one won't be defined and will be assumed zero (hence always less than
any version), and vice-versa. If we are using the Mac OS X 10.7 SDK /
iOS 5.0 SDK, then the compiler has what it needs from the
NSJSONSerialization header, so re-declaring them is an error.