Fix adding attachment to untitled doc's CouchModel
CouchAttachment's parent could only be a CouchRevision. This meant an attachment couldn't be created on an
unsaved/untitled document, because it has no revision until it's saved. Fixed this by allowing the parent of
CouchAttachment to be either a CouchDocument or CouchRevision.
Added unit tests to Test_Model.m to test adding attachments before or after the first save.

Change-Id: I84896a9ab1fda520c475e609ee0e193a838555a7