Don't post change notifications of non-winning new revisions
When CouchTouchDBDatabase gets a new-revision notification from TouchDB,  it needs to check whether
this is the winning revision of the document; otherwise it shouldn't post it or CouchDocument will get
confused. (This is another case of bug TouchDB-152, which just happens to live in CouchCocoa.)

Note: This is a dependency on a pretty new revision of TouchDB from earlier this week.